Created a prototype / example of the dictionary data format
It would be good to contain all of the raw dictionary data in plain text
files so that it can be edited easily and take advantage of source
control easily. A format like JSON or XML would involve tag names or
dictionary names, which would end up being repeated a lot over an entire
dictionary, and would be annoying to edit. A special plain text format
minimises inconvenient syntax markers.